+++ src/common/ConnectWalletBtn.vue
@@ -9,19 +9,37 @@
 </template>
 <script lang="ts" setup>
 import { AppButton } from '@/common'
-import { useWeb3ProvidersStore } from '@/store'
+import { useWeb3ProvidersStore, useChainsStore } from '@/store'
 import { computed } from 'vue'
-import { cropAddress, ErrorHandler } from '@/helpers'
+import {
+  cropAddress,
+  ErrorHandler,
+  sleep,
+  switchNetwork,
+  isMobile,
+} from '@/helpers'
 import { useI18n } from 'vue-i18n'
 import { errors } from '@/errors'
 
 const { provider } = useWeb3ProvidersStore()
+const chainStore = useChainsStore()
+
 const { t } = useI18n({ useScope: 'global' })
 
 const connect = async () => {
   try {
     if (provider.currentProvider) {
       await provider.connect()
+      await sleep(1000)
+      if (chainStore.selectedChain) {
+        await switchNetwork(chainStore.selectedChain!)
+      }
+    } else if (isMobile()) {
+      window.open(
+        `https://metamask.app.link/dapp/${window.location.host}`,
+        '_blank',
+        'noopener',
+      )
     } else {
       throw new errors.MetamaskProviderNotFound()
     }