+++ src/common/TokenSelect.vue
@@ -1,5 +1,5 @@
 <template>
-  <div :class="selectFieldClasses">
+  <div :class="selectFieldClasses" :title="errorMessage ? errorMessage : ''">
     <label v-if="label" class="token-select__label">
       {{ label }}
     </label>
@@ -79,7 +79,7 @@
               />
               <icon
                 v-else-if="option.icon"
-                class="select-field__select-dropdown-item-icon"
+                class="token-select__select-dropdown-item-icon"
                 :name="option.icon"
               />
               <span class="token-select__select-dropdown-item-text">
@@ -108,6 +108,7 @@ import {
 } from '@/helpers'
 import debounce from 'lodash/debounce'
 import { ICON_NAMES } from '@/enums'
+import { useI18n } from 'vue-i18n'
 
 type SIZES = 'big' | 'default'
 
@@ -127,6 +128,9 @@ const props = withDefaults(
     isHeadIconShown?: boolean
     size?: SIZES
     emitEmptyValueOnStartSearch?: boolean
+    isEmitSearchValueOnInput?: boolean
+    isCanSearchToken?: boolean
+    isSelectAllOptionShown?: boolean
   }>(),
   {
     valueOptions: () => [],
@@ -137,6 +141,9 @@ const props = withDefaults(
     rpcUrl: '',
     isHeadIconShown: false,
     emitEmptyValueOnStartSearch: false,
+    isEmitSearchValueOnInput: false,
+    isCanSearchToken: false,
+    isSelectAllOptionShown: false,
     size: 'default',
   },
 )
@@ -147,6 +154,7 @@ const emit = defineEmits<{
 }>()
 
 const attrs = useAttrs()
+const { t } = useI18n({ useScope: 'global' })
 
 const selectElement = ref<HTMLDivElement>()
 const searchFieldRef = ref<typeof InputField>()
@@ -176,11 +184,11 @@ const isReadonly = computed(() =>
 const headLabel = computed(() => {
   return isEthAddress(searchValue.value)
     ? cropAddress(searchValue.value, 3, 4)
-    : searchValue.value
+    : searchValue.value || selectedOption.value?.label || props.placeholder
 })
 
 const selectedOption = computed(() =>
-  props.valueOptions.find(
+  optionList.value.find(
     i =>
       i.value.toLowerCase() === searchValue.value.toLowerCase() ||
       i.label.toLowerCase() === searchValue.value.toLowerCase(),
@@ -203,15 +211,19 @@ const toggleDropdown = () => {
 const openDropdown = () => {
   if (isDisabled.value || isReadonly.value) return
   isDropdownOpen.value = true
-  isInputShown.value = true
-  nextTick(() => {
-    searchFieldRef.value?.inputRef.focus()
-  })
+  if (props.isCanSearchToken) {
+    isInputShown.value = true
+    nextTick(() => {
+      searchFieldRef.value?.inputRef.focus()
+    })
+  }
 }
 
 const closeDropdown = () => {
   isDropdownOpen.value = false
-  isInputShown.value = false
+  if (props.isCanSearchToken) {
+    isInputShown.value = false
+  }
 }
 
 const select = (value: string) => {
@@ -222,9 +234,23 @@ const select = (value: string) => {
 const getOptionList = () => {
   const regex = new RegExp(searchValue.value.trim(), 'i')
 
-  optionList.value = searchValue.value
-    ? props.valueOptions.filter(i => regex.test(i.label) || regex.test(i.value))
-    : props.valueOptions
+  optionList.value =
+    searchValue.value && props.isCanSearchToken
+      ? props.valueOptions.filter(
+          i => regex.test(i.label) || regex.test(i.value),
+        )
+      : [
+          ...(props.isSelectAllOptionShown
+            ? [
+                {
+                  label: t('token-select.all-lbl'),
+                  value: '',
+                  icon: ICON_NAMES.user,
+                },
+              ]
+            : []),
+          ...props.valueOptions,
+        ]
 }
 
 const handleSearch = async () => {
@@ -235,6 +261,8 @@ const handleSearch = async () => {
     let address = ''
     if (selectedOption.value) {
       address = selectedOption?.value?.value
+    } else if (props.isEmitSearchValueOnInput) {
+      address = searchValue.value
     } else if (isEthAddress(searchValue.value)) {
       if (!props.rpcUrl) return
       const data = await loadTokenInfo(props.rpcUrl, searchValue.value)
@@ -330,6 +358,7 @@ watch(
   height: 100%;
   min-height: toRem(24);
   position: relative;
+  cursor: pointer;
 
   @include field-text;
 
@@ -353,13 +382,6 @@ watch(
     color: var(--field-error);
     -webkit-text-fill-color: var(--field-error);
   }
-}
-
-.token-select__placeholder {
-  font: inherit;
-  opacity: 0.25;
-
-  @include field-placeholder;
 }
 
 .token-select__select-head-indicator {