+++ src/forms/create-order-form/CreateOrderFormTokensStep.vue
@@ -15,7 +15,6 @@
         placeholder="0.0000"
         :error-message="getFieldErrorMessage('amountSell')"
         :disabled="isDisabled"
-        :is-error-message-shown="false"
         @blur="touchField('amountSell')"
       />
       <input-field
@@ -24,7 +23,6 @@
         class="create-order-form-tokens-step__input"
         placeholder="0.0000"
         :error-message="getFieldErrorMessage('amountBuy')"
-        :is-error-message-shown="false"
         :disabled="isDisabled"
         @blur="touchField('amountBuy')"
       />
@@ -61,10 +59,33 @@
             @blur="touchField('tokenSell')"
           />
         </div>
+
+        <field-error-message
+          :error-message="
+            (!isSmallWidth && amountSellErrorMessage) || tokenSellErrorMessage
+          "
+        />
+
         <p class="create-order-form-tokens-step__token-network">
           {{
             $t('create-order-form-tokens-step.token-network', {
               network: networkSell?.name,
+            })
+          }}
+        </p>
+
+        <p
+          class="create-order-form-tokens-step__token-balance"
+          :class="{
+            'create-order-form-tokens-step__token-balance--loading':
+              isBalanceLoading,
+          }"
+        >
+          {{
+            $t('create-order-form-tokens-step.token-balance', {
+              balance: isBalanceLoading
+                ? $t('create-order-form-tokens-step.token-balance-loading-lbl')
+                : formatWeiAmount(tokenSellBalance, tokenSellInfo.decimals),
             })
           }}
         </p>
@@ -100,6 +121,11 @@
             @blur="touchField('tokenBuy')"
           />
         </div>
+        <field-error-message
+          :error-message="
+            (!isSmallWidth && amountBuyErrorMessage) || tokenBuyErrorMessage
+          "
+        />
         <p class="create-order-form-tokens-step__token-network">
           {{
             $t('create-order-form-tokens-step.token-network', {
@@ -136,18 +162,35 @@
 </template>
 
 <script lang="ts" setup>
-import { AppButton, AutomaticClaimBlock, TokenSelect } from '@/common'
+import {
+  AppButton,
+  TokenSelect,
+  FieldErrorMessage,
+  AutomaticClaimBlock,
+} from '@/common'
 import { InputField } from '@/fields'
 import { useFormValidation } from '@/composables'
-import { UseCreateOrderForm } from '@/types'
-import { computed, toRefs } from 'vue'
-import { required } from '@/validators'
+import { TokenInfo, UseCreateOrderForm } from '@/types'
+import { computed, reactive, ref, toRef, toRefs, watch } from 'vue'
+import {
+  required,
+  amount,
+  sameTokenInSameNetwork,
+  maxValue,
+} from '@/validators'
 import { useWindowSize } from '@vueuse/core'
 import { WINDOW_BREAKPOINTS } from '@/enums'
-import { useTokensStore } from '@/store'
+import { useTokensStore, useWeb3ProvidersStore } from '@/store'
+import {
+  loadTokenBalance,
+  getTokenInfo,
+  formatWeiAmount,
+  ErrorHandler,
+} from '@/helpers'
 
 const { width: windowWidth } = useWindowSize()
-const { tokensByChainId } = useTokensStore()
+const { tokensByChainId, tokenByAddressAndChainId } = useTokensStore()
+const { provider } = useWeb3ProvidersStore()
 
 const props = defineProps<{
   former: UseCreateOrderForm
@@ -159,9 +202,20 @@ const emit = defineEmits<{
   (e: 'next'): void
 }>()
 
+const tokenSellBalance = ref('0')
+const isBalanceLoading = ref(false)
+let tokenSellInfo = reactive<TokenInfo>({} as TokenInfo)
+
 const isSmallWidth = computed(
   () => windowWidth.value < WINDOW_BREAKPOINTS.small,
 )
+const amountSellErrorMessage = computed(() =>
+  getFieldErrorMessage('amountSell'),
+)
+const tokenSellErrorMessage = computed(() => getFieldErrorMessage('tokenSell'))
+
+const amountBuyErrorMessage = computed(() => getFieldErrorMessage('amountBuy'))
+const tokenBuyErrorMessage = computed(() => getFieldErrorMessage('tokenBuy'))
 
 const { form, networkBuy, networkSell } = toRefs(props.former)
 const tokensSell = computed(() =>
@@ -180,20 +234,86 @@ const tokensBuy = computed(() =>
   })),
 )
 
+const tokenSell = computed(() =>
+  tokenByAddressAndChainId(form.value.tokenSell, networkSell.value!.id!),
+)
+const tokenBuy = computed(() =>
+  tokenByAddressAndChainId(form.value.tokenBuy, networkBuy.value!.id!),
+)
+
+const rules = computed(() => ({
+  amountSell: {
+    required,
+    amount,
+    ...(tokenSell.value?.chain.max_amount && {
+      maxValue: maxValue(tokenSell.value?.chain.max_amount),
+    }),
+  },
+  amountBuy: {
+    required,
+    amount,
+    ...(tokenBuy.value?.chain.max_amount && {
+      maxValue: maxValue(tokenBuy.value?.chain.max_amount),
+    }),
+  },
+  tokenSell: {
+    required,
+    sameTokenInSameNetwork: sameTokenInSameNetwork(
+      networkBuy.value!,
+      networkSell.value!,
+      toRef(form.value, 'tokenBuy'),
+    ),
+  },
+  tokenBuy: {
+    required,
+    sameTokenInSameNetwork: sameTokenInSameNetwork(
+      networkBuy.value!,
+      networkSell.value!,
+      toRef(form.value, 'tokenSell'),
+    ),
+  },
+}))
+
 const { isFormValid, getFieldErrorMessage, touchField } = useFormValidation(
   form,
-  {
-    amountSell: { required },
-    amountBuy: { required },
-    tokenSell: { required },
-    tokenBuy: { required },
-  },
+  rules,
 )
 
 const handleNext = () => {
   if (!isFormValid()) return
   emit('next')
 }
+
+const loadTokenInfo = async () => {
+  if (!form.value.tokenSell) {
+    tokenSellBalance.value = '0'
+    return
+  }
+  isBalanceLoading.value = true
+  try {
+    const [balance, tokenInfo] = await Promise.all([
+      loadTokenBalance(
+        networkSell.value?.chain_params.rpc!,
+        form.value.tokenSell,
+      ),
+      getTokenInfo(networkSell.value!, form.value.tokenSell),
+    ])
+    tokenSellBalance.value = balance
+    tokenSellInfo = tokenInfo
+  } catch (e) {
+    tokenSellBalance.value = '0'
+    ErrorHandler.processWithoutFeedback(e)
+  }
+  isBalanceLoading.value = false
+}
+
+watch(
+  () => [form.value.tokenSell, provider.selectedAddress],
+  () => {
+    loadTokenInfo()
+  },
+  { immediate: true },
+)
 </script>
 
 <style lang="scss" scoped>
@@ -203,6 +323,7 @@ const handleNext = () => {
   align-items: center;
   max-width: toRem(564);
   width: 100%;
+  min-height: toRem(327);
 }
 
 .create-order-form-tokens-step__title {
@@ -229,6 +350,8 @@ const handleNext = () => {
   width: 100%;
   gap: toRem(32);
   max-width: toRem(366);
+  flex: 1;
+  align-items: flex-end;
 
   @include respond-to(tablet) {
     margin-top: toRem(12);
@@ -253,10 +376,25 @@ const handleNext = () => {
   color: var(--text-primary-main);
 }
 
+.create-order-form-tokens-step__token-balance {
+  font-size: toRem(12);
+  line-height: 1.2;
+  color: var(--text-primary-main);
+  margin-top: toRem(4);
+
+  &--loading {
+    &:after {
+      content: ' .';
+      animation: dots 1.2s steps(5, end) infinite;
+    }
+  }
+}
+
 .create-order-form-tokens-step__token-network {
   font-size: toRem(14);
   line-height: 1.2;
   color: var(--text-primary-dark);
+  margin-top: toRem(8);
 
   @include respond-to(tablet) {
     font-size: toRem(12);
@@ -291,7 +429,7 @@ const handleNext = () => {
   align-items: center;
   grid-template-columns: toRem(60) minmax(toRem(100), 1fr);
   width: calc(100% + #{toRem(3)});
-  margin: toRem(8) 0;
+  margin-top: toRem(8);
   gap: toRem(12);
 
   @include respond-to(tablet) {
@@ -323,5 +461,28 @@ const handleNext = () => {
   margin-top: toRem(26);
   padding-top: toRem(16);
   border-top: toRem(1) dashed var(--primary-main);
+}
+
+@keyframes dots {
+  0%,
+  20% {
+    color: transparent;
+    text-shadow: 0.25em 0 0 transparent, 0.5em 0 0 transparent;
+  }
+
+  40% {
+    color: var(--text-primary-dark);
+    text-shadow: 0.25em 0 0 transparent, 0.5em 0 0 transparent;
+  }
+
+  60% {
+    text-shadow: 0.25em 0 0 var(--text-primary-dark), 0.5em 0 0 transparent;
+  }
+
+  80%,
+  100% {
+    text-shadow: 0.25em 0 0 var(--text-primary-dark),
+      0.5em 0 0 var(--text-primary-dark);
+  }
 }
 </style>