+++ src/pages/Dashboard/DashboardOrderList.vue
@@ -84,10 +84,10 @@ import { useWindowSize } from '@vueuse/core'
 import { callers } from '@/api'
 
 const props = defineProps<{
-  networkSell: ChainResposne
-  matchNetwork: ChainResposne
-  tokenBuy: string
-  tokenSell: string
+  networkSell?: ChainResposne
+  matchNetwork?: ChainResposne
+  tokenBuy?: string
+  tokenSell?: string
   isSubmitting?: boolean
 }>()
 
@@ -118,10 +118,15 @@ const loadList = async () => {
       '/integrations/order-aggregator/orders',
       {
         params: {
-          'filter[src_chain]': props.networkSell.chain_params.chain_id,
-          'filter[destination_chain]': props.matchNetwork.chain_params.chain_id,
-          'filter[token_to_buy]': props.tokenBuy,
-          'filter[token_to_sell]': props.tokenSell,
+          ...(props.networkSell && {
+            'filter[src_chain]': props.networkSell.chain_params.chain_id,
+          }),
+          ...(props.matchNetwork && {
+            'filter[destination_chain]':
+              props.matchNetwork.chain_params.chain_id,
+          }),
+          ...(props.tokenBuy && { 'filter[token_to_buy]': props.tokenBuy }),
+          ...(props.tokenSell && { 'filter[token_to_sell]': props.tokenSell }),
           'filter[state]': OrderStatus.awaitingMatch,
           'page[limit]': PAGE_LIMIT,
           'page[number]': currentPage.value - 1,
@@ -149,15 +154,15 @@ Bus.on(Bus.eventList.orderMatched, () => {
 })
 
 watch(
-  () => [props.tokenSell, props.tokenBuy],
-  val => {
-    if (val[0] && val[1]) {
-      currentPage.value = 1
-      loadList()
-    } else {
-      list.value = []
-      isLoaded.value = true
-    }
+  () => [
+    props.tokenSell,
+    props.tokenBuy,
+    props.networkSell,
+    props.matchNetwork,
+  ],
+  () => {
+    currentPage.value = 1
+    loadList()
   },
   {
     immediate: true,