+++ src/types/api.types.ts
@@ -20,6 +20,8 @@ type ChainParams = {
   native_decimals: number
   native_symbol: string
   rpc: string
+  public_rpc: string
+  chain_type: string
 }
 
 export type ChainResposne = JsonApiRecordBase<'chain'> & {
@@ -48,13 +50,13 @@ export type TokenChain = {
   token_id: string
   token_type: string
   contract_address: string
+  max_amount: string
 }
 
 export type TokenResponse = JsonApiRecordBase<'token'> & {
   icon: string
   name: string
   symbol: string
-  decimals: number
   token_type: string
   chains: TokenChain[]
 }
@@ -63,7 +65,6 @@ export type Token = JsonApiRecordBase<'token'> & {
   icon: string
   name: string
   symbol: string
-  decimals: number
   token_type: string
   chain: TokenChain
 }