+++ CHANGELOG.md
@@ -7,6 +7,77 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+## [1.0.0-rc.24] - 2023-05-17
+#### Fixed
+- Typo
+
+## [1.0.0-rc.23] - 2023-04-27
+#### Added
+- Toggle network on dashboard and create an order form
+
+## [1.0.0-rc.22] - 2023-04-26
+#### Fixed
+- A bug with a freeze on scroll
+
+## [1.0.0-rc.21] - 2023-04-26
+#### Added
+- Show all orders on dashboard
+- Animating network in table
+
+## [1.0.0-rc.20] - 2023-04-21
+#### Fixed
+- Error messages
+
+## [1.0.0-rc.19] - 2023-04-20
+#### Added
+- Error message
+
+## [1.0.0-rc.18] - 2023-04-12
+#### Added
+- Max amount for tokens
+
+## [1.0.0-rc.17] - 2023-04-11
+#### Changed
+- The ability to search for tokens is blocked
+
+## [1.0.0-rc.16] - 2023-04-10
+#### Added
+- All option to select field
+
+#### Fixed
+- A bug with switch chain in metamask
+
+## [1.0.0-rc.15] - 2023-04-07
+#### Changed
+- Meta tags
+
+## [1.0.0-rc.14] - 2023-04-06
+#### Fixed
+- A bug with switch chain in metamask browser
+
+## [1.0.0-rc.13] - 2023-04-05
+#### Fixed
+- A bug with switch chain in metamask browser
+
+## [1.0.0-rc.12] - 2023-04-05
+#### Fixed
+- A bug with select chain
+
+## [1.0.0-rc.11] - 2023-04-04
+#### Added
+- Display the balance of owned token on create an order form
+
+#### Fixed
+- A bug with create order in the same networks with the same token
+
+## [1.0.0-rc.10] - 2023-04-03
+#### Fixed
+- Unuserfriendly error when user with empty balance tries to create an order
+- A bug with opportunity to create an order with 0 amount
+
+#### Removed
+- Default tokens from dashboard
+
 ## [1.0.0-rc.9] - 2023-03-01
 #### Changed
 - Getting lists from the backend instead of contracts
@@ -57,7 +128,22 @@ for further information about branching and tagging conventions.
 #### Under the hood changes
 - Initiated and setup project
 
-[Unreleased]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.9...main
+[Unreleased]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.24...main
+[1.0.0-rc.24]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.23...v1.0.0-rc.24
+[1.0.0-rc.23]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.22...v1.0.0-rc.23
+[1.0.0-rc.22]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.21...v1.0.0-rc.22
+[1.0.0-rc.21]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.20...v1.0.0-rc.21
+[1.0.0-rc.20]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.19...v1.0.0-rc.20
+[1.0.0-rc.19]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.18...v1.0.0-rc.19
+[1.0.0-rc.18]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.17...v1.0.0-rc.18
+[1.0.0-rc.17]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.16...v1.0.0-rc.17
+[1.0.0-rc.16]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.15...v1.0.0-rc.16
+[1.0.0-rc.15]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.14...v1.0.0-rc.15
+[1.0.0-rc.14]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.13...v1.0.0-rc.14
+[1.0.0-rc.13]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.12...v1.0.0-rc.13
+[1.0.0-rc.12]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.11...v1.0.0-rc.12
+[1.0.0-rc.11]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.10...v1.0.0-rc.11
+[1.0.0-rc.10]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.9...v1.0.0-rc.10
 [1.0.0-rc.9]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.8...v1.0.0-rc.9
 [1.0.0-rc.8]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.7...v1.0.0-rc.8
 [1.0.0-rc.7]: https://github.com/Swapica/web-client/compare/v1.0.0-rc.6...v1.0.0-rc.7